connをインスタンス変数化。memo_db.rb作成。deallocate_statementメソッド追加
